update: prod fixes
