+++ packages/gofamintui-app/prisma/schema.prisma
@@ -8,16 +8,16 @@ generator client {
   provider = "prisma-client-js"
 }
 
-datasource db {
-  provider  = "postgresql"
-  url       = env("DATABASE_URL")
-  directUrl = env("DIRECT_URL")
-}
-        
 // datasource db {
-//   provider = "sqlite"
-//   url      = env("DATABASE_URL")
+//   provider  = "postgresql"
+//   url       = env("DATABASE_URL")
+//   directUrl = env("DIRECT_URL")
 // }
+        
+datasource db {
+  provider = "sqlite"
+  url      = env("DATABASE_URL")
+}
 
 
 model User {
