+++ packages/gofamintui-cms/schemaTypes/documents/author.ts
@@ -10,6 +10,9 @@ const APPLICATION_STATUSES: Array<{title: string; value: ApplicationStatus}> = [
 
 ]
 
+
+
+
 export const author = defineType({
   name: 'author',
   title: 'Author Details',
